Fix too harsh slowdown
Fix #14

For some reason, the position delta based bhop UPS values had some
jitter in them, such as:

`0, 261, 245, 246, 492, 242, 0, 246, 246, 246, 493, 246, 246, 0`

where some ticks saw an apparent UPS value of 0 (no positional change
from last tick), and other saw a double value to compensate for this
drift.

This commit reworks the speed calculation as an average over several
ticks (2 by default), to account for this jitter - penalties are still
applied based on real ups of that tick, but using the averaged inverse
impulse, instead.

Because this buffering essentially introduces a 1 hop delay to the
calculation, the cvar `sm_nt_anti_ghosthop_n_allowed_hops` has been
removed, and replaced by `sm_nt_anti_ghosthop_n_extra_hops` which
defaults to 0. This value of 0 has the buffered hops built-in, and is
the recommended value.